BUG: Inconsistent ordering of table columns.
When using d3 to store the json table object into
a csv file, the order of the columns were not maintained.

This causes the VTKTanglegramItem rendering failure with the
corresponding lines. The tanglegram relies on the input
correspondence table to have the first column data being a string
array.

This commit uses text strings (in stead of json object) to store
the table into the mongodb, and writes out the same text strings
to the csv file.