add support for python analyses
Add the "compare trees" analysis to the app.  This involved
adding support for python analyses in addition to the R analyses
that we already support.  Compare trees also involved some minor
changes to the tree_heatmap visualization.  It now checks if the
tree has a vertex array named "property.differences".  If so,
this array is used to color the edges of the tree.